Convert 4 CK fusion solvers from direct CK linking to dlopen runtime loading
Refactor ConvCKIgemmFwdBiasActivFused, ConvCKIgemmFwdBiasResAddActivFused,
ConvCKIgemmGrpFwdActivFused, and ConvCKIgemmGrpFwdBiasActivFused to use
CKGroupedConvLibLoader instead of compile-time CK template instantiation.

This extends the dlopen infrastructure from PR #5720 to cover the remaining
4 fused CK solvers, enabling multi-arch support without recompiling MIOpen.

Changes:
- Extend CKSolverType enum with 4 new entries (FusedBiasActiv,
  FusedBiasResAddActiv, FusedGrpActiv, FusedGrpBiasActiv)
- Add 20 extern "C" interface declarations (5 per solver)
- Create 4 ck_impl files implementing the extern "C" functions
- Wire solver bindings in ck_grouped_conv_lib_loader.cpp
- Rewrite 4 solver files to delegate to CKGroupedConvLibLoader
- Bump CK_GROUPED_CONV_API_VERSION from 2 to 3
- Add unit test coverage for all 4 new solver types in loader tests

Net: +2142/-1803 lines across 13 files.
