+++ projects/miopen/src/solver/conv_ck_igemm_fwd_bias_activ_fused.cpp
@@ -28,296 +28,46 @@
 #include <vector>
 #include <cstdint>
 
-#include <miopen/check_numerics.hpp>
 #include <miopen/env.hpp>
 #include <miopen/fusion/solvers.hpp>
 #include <miopen/generic_search.hpp>
 #include <miopen/conv/data_invoke_params.hpp>
-#include <miopen/solver/problem_description_interpreter.hpp>
-#if MIOPEN_BACKEND_HIP && MIOPEN_USE_COMPOSABLEKERNEL
-#include <ck/tensor_operation/gpu/device/device_conv_fwd_bias_activation.hpp>
-#endif
+#include <miopen/solver/ck_grouped_conv_lib_loader.hpp>
+
 MIOPEN_DECLARE_ENV_VAR_BOOL(MIOPEN_DEBUG_CONV_CK_IGEMM_FWD_BIAS_ACTIV)
-
-#if MIOPEN_BACKEND_HIP && MIOPEN_USE_COMPOSABLEKERNEL
-// Forward declare CK's function.
-namespace ck {
-namespace tensor_operation {
-namespace device {
-using DeviceConvFwdBiasReluPtr = ck::tensor_operation::device::DeviceConvFwdBiasActivationPtr<
-    ck::tensor_operation::element_wise::PassThrough,
-    ck::tensor_operation::element_wise::PassThrough,
-    ck::tensor_operation::element_wise::AddRelu>;
-
-namespace instance {
-
-void add_device_conv2d_fwd_xdl_c_shuffle_bias_relu_nhwc_kyxc_nhwk_f16_instances(
-    std::vector<DeviceConvFwdBiasReluPtr>&);
-
-} // namespace instance
-} // namespace device
-} // namespace tensor_operation
-} // namespace ck
-#endif
 
 namespace miopen {
 namespace solver {
 namespace fusion {
 
-#if MIOPEN_BACKEND_HIP && MIOPEN_USE_COMPOSABLEKERNEL
-namespace {
-
-struct CKArgs
-{
-    CKArgs(const miopen::conv::ProblemDescription& problem)
-    {
-        N        = ProblemInterpreter::GetBatchN(problem);
-        K        = ProblemInterpreter::GetOutputChannelK(problem);
-        C        = ProblemInterpreter::GetInputChannelC(problem);
-        input    = {ProblemInterpreter::GetInputHeightHi(problem),
-                    ProblemInterpreter::GetInputWidthWi(problem)};
-        output   = {ProblemInterpreter::GetOutputHeightHo(problem),
-                    ProblemInterpreter::GetOutputWidthWo(problem)};
-        filter   = {ProblemInterpreter::GetFilterHeightY(problem),
-                    ProblemInterpreter::GetFilterWidthX(problem)};
-        strides  = {ProblemInterpreter::GetAdjustedConvolutionStrideH(problem),
-                    ProblemInterpreter::GetAdjustedConvolutionStrideW(problem)};
-        dilation = {ProblemInterpreter::GetAdjustedConvolutionDilationH(problem),
-                    ProblemInterpreter::GetAdjustedConvolutionDilationW(problem)};
-        lPadding = {ProblemInterpreter::GetInputLeftPadH(problem),
-                    ProblemInterpreter::GetInputLeftPadW(problem)};
-        rPadding = {ProblemInterpreter::GetAdjustedInputRightPadH(problem),
-                    ProblemInterpreter::GetAdjustedInputRightPadW(problem)};
-    }
-    int N;
-    int K;
-    int C;
-    std::vector<int> input;
-    std::vector<int> output;
-    std::vector<int> filter;
-    std::vector<int> strides;
-    std::vector<int> dilation;
-    std::vector<int> lPadding;
-    std::vector<int> rPadding;
-};
-
-} // namespace
-
-template <typename DataType>
-void PerformanceConfigConvCKIgemmFwdBiasActivFused::Init(
-    const miopen::conv::ProblemDescription& problem)
-{
-    const auto& args = CKArgs{problem};
-    std::vector<ck::tensor_operation::device::DeviceConvFwdBiasReluPtr> conv_ptrs;
-    ck::tensor_operation::device::instance::
-        add_device_conv2d_fwd_xdl_c_shuffle_bias_relu_nhwc_kyxc_nhwk_f16_instances(conv_ptrs);
-    assert(!conv_ptrs.empty());
-    for(const auto& it : conv_ptrs)
-    {
-        auto argument_ptr = it->MakeArgumentPointer(nullptr,
-                                                    nullptr,
-                                                    nullptr,
-                                                    nullptr,
-                                                    args.N,
-                                                    args.K,
-                                                    args.C,
-                                                    args.input,
-                                                    args.filter,
-                                                    args.output,
-                                                    args.strides,
-                                                    args.dilation,
-                                                    args.lPadding,
-                                                    args.rPadding,
-                                                    {},
-                                                    {},
-                                                    {});
-        if(it->IsSupportedArgument(argument_ptr.get()))
-        {
-            valid_kernels.push_back(it->GetTypeString());
-        }
-    }
-
-    assert(!valid_kernels.empty());
-    this->index     = 0;
-    this->kernel_id = valid_kernels[0];
-}
-
-template <typename DataType>
-bool PerformanceConfigConvCKIgemmFwdBiasActivFused::CheckIsSupportCKArgs(
-    const miopen::conv::ProblemDescription& problem) const
-{
-    const auto& args = CKArgs{problem};
-    std::vector<ck::tensor_operation::device::DeviceConvFwdBiasReluPtr> conv_ptrs;
-    ck::tensor_operation::device::instance::
-        add_device_conv2d_fwd_xdl_c_shuffle_bias_relu_nhwc_kyxc_nhwk_f16_instances(conv_ptrs);
-
-    int i = 0;
-    for(; i < conv_ptrs.size(); i++)
-    {
-        if(conv_ptrs[i]->GetTypeString() == this->kernel_id)
-        {
-            break;
-        }
-    }
-    if(i == valid_kernels.size())
-    {
-        return false;
-    }
-    auto argument_ptr = conv_ptrs[i]->MakeArgumentPointer(nullptr,
-                                                          nullptr,
-                                                          nullptr,
-                                                          nullptr,
-                                                          args.N,
-                                                          args.K,
-                                                          args.C,
-                                                          args.input,
-                                                          args.filter,
-                                                          args.output,
-                                                          args.strides,
-                                                          args.dilation,
-                                                          args.lPadding,
-                                                          args.rPadding,
-                                                          {},
-                                                          {},
-                                                          {});
-    return conv_ptrs[i]->IsSupportedArgument(argument_ptr.get());
-}
-
-template <typename DataType>
-bool ConvCKIgemmFwdBiasActivFused::CheckCKApplicability(
-    const miopen::conv::ProblemDescription& problem) const
-{
-    std::vector<ck::tensor_operation::device::DeviceConvFwdBiasReluPtr> conv_ptrs;
-    ck::tensor_operation::device::instance::
-        add_device_conv2d_fwd_xdl_c_shuffle_bias_relu_nhwc_kyxc_nhwk_f16_instances(conv_ptrs);
-    assert(!conv_ptrs.empty());
-    const auto& args = CKArgs{problem};
-    for(const auto& it : conv_ptrs)
-    {
-        auto argument_ptr = it->MakeArgumentPointer(nullptr,
-                                                    nullptr,
-                                                    nullptr,
-                                                    nullptr,
-                                                    args.N,
-                                                    args.K,
-                                                    args.C,
-                                                    args.input,
-                                                    args.filter,
-                                                    args.output,
-                                                    args.strides,
-                                                    args.dilation,
-                                                    args.lPadding,
-                                                    args.rPadding,
-                                                    {},
-                                                    {},
-                                                    {});
-        if(it->IsSupportedArgument(argument_ptr.get()))
-            return true;
-    }
-    return false;
-}
-
-namespace {
-
-template <typename DataType>
-void RunCKSolution(const Handle& handle,
-                   const AnyInvokeParams& primitive_parameters,
-                   const miopen::conv::ProblemDescription& problem,
-                   const PerformanceConfigConvCKIgemmFwdBiasActivFused& config)
-{
-    const auto& args = CKArgs{problem};
-    std::vector<ck::tensor_operation::device::DeviceConvFwdBiasReluPtr> conv_ptrs;
-    ck::tensor_operation::device::instance::
-        add_device_conv2d_fwd_xdl_c_shuffle_bias_relu_nhwc_kyxc_nhwk_f16_instances(conv_ptrs);
-
-    int id = 0;
-    for(; id < conv_ptrs.size(); id++)
-    {
-        if(conv_ptrs[id]->GetTypeString() == config.kernel_id)
-        {
-            break;
-        }
-    }
-    assert(id < conv_ptrs.size());
-    auto& conv_ck          = conv_ptrs.at(id);
-    const auto& invoke_ctx = primitive_parameters.CastTo<miopen::fusion::FusionInvokeParams>();
-    const auto& wei_buf =
-        dynamic_cast<miopen::fusion::ConvolutionOpInvokeParam&>(*invoke_ctx.op_args.params[0])
-            .weights;
-    const auto& bias_buf =
-        dynamic_cast<miopen::fusion::BiasOpInvokeParam&>(*invoke_ctx.op_args.params[1]).bdata;
-
-    auto argument_ptr = conv_ck->MakeArgumentPointer(
-        const_cast<void*>( // NOLINT (cppcoreguidelines-pro-type-const-cast)
-            static_cast<const void*>(invoke_ctx.in)),
-        const_cast<void*>( // NOLINT (cppcoreguidelines-pro-type-const-cast)
-            static_cast<const void*>(wei_buf)),
-        invoke_ctx.out,
-        const_cast<void*>( // NOLINT (cppcoreguidelines-pro-type-const-cast)
-            static_cast<const void*>(bias_buf)),
-        args.N,
-        args.K,
-        args.C,
-        args.input,
-        args.filter,
-        args.output,
-        args.strides,
-        args.dilation,
-        args.lPadding,
-        args.rPadding,
-        {},
-        {},
-        {});
-    auto invoker_ptr            = conv_ck->MakeInvokerPointer();
-    const auto enable_profiling = handle.IsProfilingEnabled();
-
-    float elapsed_time =
-        invoker_ptr->Run(argument_ptr.get(), {handle.GetStream(), enable_profiling});
-    if(enable_profiling)
-    {
-        handle.ResetKernelTime();
-        handle.AccumKernelTime(elapsed_time);
-    }
-}
-
-} // namespace
-#endif
-
 void PerformanceConfigConvCKIgemmFwdBiasActivFused::HeuristicInit(
     const FusionDescription& fdesc_problem)
 {
-#if !MIOPEN_BACKEND_HIP || !MIOPEN_USE_COMPOSABLEKERNEL
-    std::ignore = fdesc_problem;
-#else
-    const auto conv_problem = fdesc_problem.GetConvProblem(0, miopen::conv::Direction::Forward);
-    switch(conv_problem.GetInDataType())
-    {
-    case miopenHalf: Init<ck::half_t>(conv_problem); break;
-    case miopenFloat8_fnuz:
-    case miopenBFloat8_fnuz:
-    case miopenInt8:
-    case miopenFloat:
-    case miopenInt32:
-    case miopenInt64:
-    case miopenBFloat16:
-    case miopenDouble:
-    default: MIOPEN_THROW("Unsupported datatype");
-    }
+    const auto& loader = CKGroupedConvLibLoader::Get(GetCurrentDeviceName());
+    if(!loader.IsLoaded())
+        return;
 
-#endif
+    const auto conv_problem = fdesc_problem.GetConvProblem(0, miopen::conv::Direction::Forward);
+    const auto data_type    = conv_problem.GetInDataType();
+
+    valid_kernels = loader.FillValidKernels(
+        CKSolverType::FusedBiasActiv, conv_problem, data_type, false);
+
+    if(!valid_kernels.empty())
+    {
+        index     = 0;
+        kernel_id = valid_kernels[0];
+    }
 }
 
 bool PerformanceConfigConvCKIgemmFwdBiasActivFused::SetNextValue(
     const FusionDescription& fdesc_problem)
 {
-#if !MIOPEN_BACKEND_HIP || !MIOPEN_USE_COMPOSABLEKERNEL
-    std::ignore = fdesc_problem;
-    return false;
-#else
     if(this->valid_kernels.empty())
     {
         this->HeuristicInit(fdesc_problem);
-        assert(!valid_kernels.empty());
+        if(valid_kernels.empty())
+            return false;
         return true;
     }
     if((this->index + 1) < valid_kernels.size())
@@ -328,7 +78,6 @@ bool PerformanceConfigConvCKIgemmFwdBiasActivFused::SetNextValue(
     }
     else
         return false;
-#endif
 }
 
 bool PerformanceConfigConvCKIgemmFwdBiasActivFused::IsValidValue() const
@@ -339,27 +88,14 @@ bool PerformanceConfigConvCKIgemmFwdBiasActivFused::IsValidValue() const
 bool PerformanceConfigConvCKIgemmFwdBiasActivFused::IsValid(
     const FusionContext&, const FusionDescription& fdesc_problem) const
 {
-#if !MIOPEN_BACKEND_HIP || !MIOPEN_USE_COMPOSABLEKERNEL
-    std::ignore = fdesc_problem;
-    return false;
-#else
-    // Extract convolution problem from the fusion context.
+    const auto& loader = CKGroupedConvLibLoader::Get(GetCurrentDeviceName());
+    if(!loader.IsLoaded())
+        return false;
+
     const auto conv_problem = fdesc_problem.GetConvProblem(0, miopen::conv::Direction::Forward);
-    switch(conv_problem.GetInDataType())
-    {
-    case miopenHalf: return CheckIsSupportCKArgs<ck::half_t>(conv_problem);
-    case miopenFloat8_fnuz:
-    case miopenBFloat8_fnuz:
-    case miopenInt8:
-    case miopenFloat:
-    case miopenInt32:
-    case miopenInt64:
-    case miopenBFloat16:
-    case miopenDouble:
-    default: MIOPEN_THROW("Unsupported datatype");
-    }
-    return false;
-#endif
+    const auto data_type    = conv_problem.GetInDataType();
+    return loader.IsArgsSupported(
+        CKSolverType::FusedBiasActiv, conv_problem, kernel_id, data_type, false);
 }
 
 bool PerformanceConfigConvCKIgemmFwdBiasActivFused::operator==(
@@ -397,11 +133,6 @@ ConvCKIgemmFwdBiasActivFused::Search(const FusionContext& ctx,
 bool ConvCKIgemmFwdBiasActivFused::IsApplicable(const FusionContext& ctx,
                                                 const FusionDescription& fdesc_problem) const
 {
-#if !MIOPEN_BACKEND_HIP || !MIOPEN_USE_COMPOSABLEKERNEL
-    std::ignore = ctx;
-    std::ignore = fdesc_problem;
-    return false;
-#else
     const auto& desc = *fdesc_problem.fusion_plan_desc;
     if(desc.op_map.empty())
     {
@@ -445,57 +176,26 @@ bool ConvCKIgemmFwdBiasActivFused::IsApplicable(const FusionContext& ctx,
     if(conv_problem.GetGroupCount() > 1)
         return false;
 
-    switch(conv_problem.GetInDataType())
-    {
-    case miopenHalf: return CheckCKApplicability<ck::half_t>(conv_problem);
-    case miopenFloat8_fnuz:
-    case miopenBFloat8_fnuz:
-    case miopenInt8:
-    case miopenFloat:
-    case miopenInt32:
-    case miopenInt64:
-    case miopenBFloat16:
-    case miopenDouble:
-    default: MIOPEN_THROW("Unsupported datatype");
-    }
-    return false;
-#endif
+    const auto& loader = CKGroupedConvLibLoader::Get(arch);
+    if(!loader.IsLoaded())
+        return false;
+
+    return loader.IsApplicable(
+        CKSolverType::FusedBiasActiv, conv_problem, conv_problem.GetInDataType(), false);
 }
 
 ConvSolution ConvCKIgemmFwdBiasActivFused::GetSolution(
-    const FusionContext&,
+    const FusionContext& ctx,
     const FusionDescription& fdesc_problem,
     const PerformanceConfigConvCKIgemmFwdBiasActivFused& config) const
 {
-#if !MIOPEN_BACKEND_HIP || !MIOPEN_USE_COMPOSABLEKERNEL
-    std::ignore = fdesc_problem;
-    std::ignore = config;
-    return {};
-#else
+    const auto& loader = CKGroupedConvLibLoader::Get(ctx.GetStream().GetDeviceName());
+    if(!loader.IsLoaded())
+        return {};
+
     const auto conv_problem = fdesc_problem.GetConvProblem(0, miopen::conv::Direction::Forward);
-    ConvSolution result;
-    result.invoker_factory = [=](const std::vector<Kernel>& kernels) {
-        std::ignore = kernels;
-        return [=](const Handle& handle, const AnyInvokeParams& primitive_parameters) {
-            switch(conv_problem.GetInDataType())
-            {
-            case miopenHalf:
-                RunCKSolution<ck::half_t>(handle, primitive_parameters, conv_problem, config);
-                break;
-            case miopenFloat8_fnuz:
-            case miopenBFloat8_fnuz:
-            case miopenInt8:
-            case miopenFloat:
-            case miopenInt32:
-            case miopenInt64:
-            case miopenBFloat16:
-            case miopenDouble:
-            default: MIOPEN_THROW("Unsupported datatype");
-            }
-        };
-    };
-    return result;
-#endif
+    return loader.GetSolution(
+        CKSolverType::FusedBiasActiv, ctx, conv_problem, config.kernel_id, false);
 }
 
 } // namespace fusion
