+++ projects/miopen/test/gtest/unit_ck_grouped_conv_loader.cpp
@@ -367,6 +367,56 @@ TEST(CPU_CKGroupedConvLoader_NONE, LoaderReturnsEmptyOnFailure)
     EXPECT_EQ(loader.GetWorkspaceSize(CKSolverType::GrpConv3dWrw, problem, miopenHalf, false), 0u);
     EXPECT_EQ(loader.GetSolution(CKSolverType::GrpConv3dWrw, ctx, problem, "dummy", false).status,
               miopenStatusInternalError);
+
+    EXPECT_TRUE(
+        loader.FillValidKernels(CKSolverType::FusedBiasActiv, problem, miopenHalf, false).empty());
+    EXPECT_FALSE(loader.IsApplicable(CKSolverType::FusedBiasActiv, problem, miopenHalf, false));
+    EXPECT_FALSE(loader.IsArgsSupported(
+        CKSolverType::FusedBiasActiv, problem, "dummy_kernel", miopenHalf, false));
+    EXPECT_EQ(loader.GetWorkspaceSize(CKSolverType::FusedBiasActiv, problem, miopenHalf, false),
+              0u);
+    EXPECT_EQ(
+        loader.GetSolution(CKSolverType::FusedBiasActiv, ctx, problem, "dummy", false).status,
+        miopenStatusInternalError);
+
+    EXPECT_TRUE(
+        loader.FillValidKernels(CKSolverType::FusedBiasResAddActiv, problem, miopenHalf, false)
+            .empty());
+    EXPECT_FALSE(
+        loader.IsApplicable(CKSolverType::FusedBiasResAddActiv, problem, miopenHalf, false));
+    EXPECT_FALSE(loader.IsArgsSupported(
+        CKSolverType::FusedBiasResAddActiv, problem, "dummy_kernel", miopenHalf, false));
+    EXPECT_EQ(
+        loader.GetWorkspaceSize(CKSolverType::FusedBiasResAddActiv, problem, miopenHalf, false),
+        0u);
+    EXPECT_EQ(
+        loader.GetSolution(CKSolverType::FusedBiasResAddActiv, ctx, problem, "dummy", false)
+            .status,
+        miopenStatusInternalError);
+
+    EXPECT_TRUE(
+        loader.FillValidKernels(CKSolverType::FusedGrpActiv, problem, miopenHalf, false).empty());
+    EXPECT_FALSE(loader.IsApplicable(CKSolverType::FusedGrpActiv, problem, miopenHalf, false));
+    EXPECT_FALSE(loader.IsArgsSupported(
+        CKSolverType::FusedGrpActiv, problem, "dummy_kernel", miopenHalf, false));
+    EXPECT_EQ(loader.GetWorkspaceSize(CKSolverType::FusedGrpActiv, problem, miopenHalf, false),
+              0u);
+    EXPECT_EQ(
+        loader.GetSolution(CKSolverType::FusedGrpActiv, ctx, problem, "dummy", false).status,
+        miopenStatusInternalError);
+
+    EXPECT_TRUE(
+        loader.FillValidKernels(CKSolverType::FusedGrpBiasActiv, problem, miopenHalf, false)
+            .empty());
+    EXPECT_FALSE(
+        loader.IsApplicable(CKSolverType::FusedGrpBiasActiv, problem, miopenHalf, false));
+    EXPECT_FALSE(loader.IsArgsSupported(
+        CKSolverType::FusedGrpBiasActiv, problem, "dummy_kernel", miopenHalf, false));
+    EXPECT_EQ(
+        loader.GetWorkspaceSize(CKSolverType::FusedGrpBiasActiv, problem, miopenHalf, false), 0u);
+    EXPECT_EQ(
+        loader.GetSolution(CKSolverType::FusedGrpBiasActiv, ctx, problem, "dummy", false).status,
+        miopenStatusInternalError);
 }
 
 TEST(CPU_CKGroupedConvLoader_NONE, IsDeterministicSplitKValid)
